# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse:
    properties:
      acceptance_criteria:
        type: string
      assignee_id:
        type: string
      assignee_name:
        type: string
      column_id:
        type: string
      completed_at:
        type: string
      completed_subtasks:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        type: string
      implementation_note:
        type: string
      is_done:
        type: boolean
      position:
        type: number
      priority:
        type: string
      tags:
        items:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.TagResponse'
        type: array
      title:
        type: string
      total_subtasks:
        type: integer
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateCardRequest:
    properties:
      assignee_id:
        type: string
      column_id:
        type: string
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      position:
        minimum: 0
        type: number
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      subtasks:
        items:
          type: string
        maxItems: 50
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - column_id
    - title
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MoveCardRequest:
    properties:
      column_id:
        type: string
      position:
        type: number
    required:
    - column_id
    - position
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.TagResponse:
    properties:
      board_id:
        type: string
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ToggleCardDoneRequest:
    properties:
      is_done:
        type: boolean
    required:
    - is_done
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateCardRequest:
    properties:
      acceptance_criteria:
        maxLength: 10000
        type: string
      assignee_id:
        type: string
      changed_fields:
        description: |-
          ChangedFields is the client-computed diff, recorded on the activity row so the
          feed can say what changed without re-reading the previous row.
        items:
          type: string
        maxItems: 20
        type: array
      description:
        maxLength: 5000
        type: string
      due_date:
        type: string
      estimated_hours:
        maximum: 10000
        minimum: 0
        type: number
      implementation_note:
        maxLength: 10000
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Turtask
  description: |-
    Multi-board Kanban + task management with realtime sync.
    Auth uses an HttpOnly `auth_token` cookie issued by /api/auth/login or /api/auth/oauth.
  title: Turtask API
  version: "1.0"
paths:
...
  /api/cards:
    post:
      consumes:
      - application/json
      parameters:
      - description: Column + title (+ optional fields)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create card
      tags:
      - cards
  /api/cards/{cardID}:
    delete:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete card
      tags:
      - cards
    get:
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get card detail
      tags:
      - cards
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.CardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update card
      tags:
      - cards
  /api/cards/{cardID}/done:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Desired done state
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.ToggleCardDoneRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Toggle card done
      tags:
      - cards
  /api/cards/{cardID}/move:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Card UUID
        in: path
        name: cardID
        required: true
        type: string
      - description: Target column + position
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_dto.MoveCardRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_aumputthipong_mini-erp-kanban_backend_internal_httputil.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Move card
      tags:
      - cards
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"